Co-locate useHamburgerMenu with HamburgerMenu components
